chore: use scenario for unit testing
