+++ test-requirements.in
@@ -6,3 +6,4 @@ pytest
 pytest-operator
 pytest-asyncio==0.21.2
 ruff
+ops-scenario
